+++ sdk/communication/Azure.Communication.Identity/tests/CommunicationIdentityClient/CommunicationIdentityClientLiveTests.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Linq;
 using System.Threading.Tasks;
+using Azure.Communication.Identity.Models;
 using Azure.Communication.Tests;
 using Azure.Core;
 using Azure.Core.TestFramework;
@@ -22,11 +23,16 @@ namespace Azure.Communication.Identity.Tests
     public class CommunicationIdentityClientLiveTests : CommunicationIdentityClientLiveTestBase
     {
         /// <summary>
+        /// Options used to exchange an AAD access token of a Teams user for a new Communication Identity access token.
+        /// </summary>
+        internal GetTokenForTeamsUserOptions CTEOptions;
+        /// <summary>
         /// Initializes a new instance of the <see cref="CommunicationIdentityClient"/> class.
         /// </summary>
         /// <param name="isAsync">A flag used by the Azure Core Test Framework to differentiate between tests for asynchronous and synchronous methods.</param>
         public CommunicationIdentityClientLiveTests(bool isAsync) : base(isAsync)
         {
+            CTEOptions = createTeamsUserParams().Result;
         }
 
         [Test]
@@ -137,10 +143,8 @@ namespace Azure.Communication.Identity.Tests
                 Assert.Ignore("Ignore exchange teams token test if flag is enabled.");
             }
 
-            var teamsUserParams = await createTeamsUserParams();
-
             CommunicationIdentityClient client = CreateClientWithConnectionString();
-            Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(teamsUserParams.Token, teamsUserParams.AppId, teamsUserParams.UserId);
+            Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(CTEOptions);
             Assert.IsNotNull(tokenResponse.Value);
             Assert.IsFalse(string.IsNullOrWhiteSpace(tokenResponse.Value.Token));
         }
@@ -152,14 +156,13 @@ namespace Azure.Communication.Identity.Tests
         [TestCase(true, true, true, "token", TestName = "GetTokenForTeamsUserWithNullParamsShouldThrow")]
         public async Task GetTokenForTeamsUserWithNullParamsShouldThrow(bool isTokenNull, bool isAppIdNull, bool isUserIdNull, string paramName)
         {
-            var teamsUserParams = await createTeamsUserParams();
-            var token = isTokenNull ? null : teamsUserParams.Token;
-            var appId = isAppIdNull ? null : teamsUserParams.AppId;
-            var userId = isUserIdNull ? null : teamsUserParams.UserId;
+            var teamsUserAadToken = isTokenNull ? null : CTEOptions.TeamsUserAadToken;
+            var clientId = isAppIdNull ? null : CTEOptions.ClientId;
+            var userObjectId = isUserIdNull ? null : CTEOptions.UserObjectId;
             try
             {
                 CommunicationIdentityClient client = CreateClientWithConnectionString();
-                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(token, appId, userId);
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(teamsUserAadToken, clientId, userObjectId));
             }
             catch (ArgumentNullException ex)
             {
@@ -177,8 +180,7 @@ namespace Azure.Communication.Identity.Tests
             try
             {
                 CommunicationIdentityClient client = CreateClientWithConnectionString();
-                var teamsUserParams = await createTeamsUserParams();
-                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(token, teamsUserParams.AppId, teamsUserParams.UserId);
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(token, CTEOptions.ClientId, CTEOptions.UserObjectId));
             }
             catch (RequestFailedException ex)
             {
@@ -196,8 +198,7 @@ namespace Azure.Communication.Identity.Tests
             try
             {
                 CommunicationIdentityClient client = CreateClientWithConnectionString();
-                var teamsUserParams = await createTeamsUserParams();
-                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(TestEnvironment.CommunicationExpiredTeamsToken, teamsUserParams.AppId, teamsUserParams.UserId);
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(TestEnvironment.CommunicationExpiredTeamsToken, CTEOptions.ClientId, CTEOptions.UserObjectId));
             }
             catch (RequestFailedException ex)
             {
@@ -210,20 +211,19 @@ namespace Azure.Communication.Identity.Tests
         }
 
         [Test]
-        [TestCase("", TestName = "GetTokenForTeamsUserWithEmptyAppIdShouldThrow")]
-        [TestCase("invalid", TestName = "GetTokenForTeamsUserWithInvalidAppIdShouldThrow")]
-        public async Task GetTokenForTeamsUserWithInvalidAppIdShouldThrow(string appId)
+        [TestCase("", TestName = "GetTokenForTeamsUserWithEmptyClientIdShouldThrow")]
+        [TestCase("invalid", TestName = "GetTokenForTeamsUserWithInvalidClientIdShouldThrow")]
+        public async Task GetTokenForTeamsUserWithInvalidClientIdShouldThrow(string clientId)
         {
             if (TestEnvironment.ShouldIgnoreIdentityExchangeTokenTest)
             {
                 Assert.Ignore("Ignore exchange teams token test if flag is enabled.");
             }
 
-            var teamsUserParams = await createTeamsUserParams();
             try
             {
                 CommunicationIdentityClient client = CreateClientWithConnectionString();
-                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(teamsUserParams.Token, appId, teamsUserParams.UserId);
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(CTEOptions.TeamsUserAadToken, clientId, CTEOptions.UserObjectId));
             }
             catch (RequestFailedException ex)
             {
@@ -236,18 +236,17 @@ namespace Azure.Communication.Identity.Tests
         }
 
         [Test]
-        public async Task GetTokenForTeamsUserWithWrongAppIdShouldThrow()
+        public async Task GetTokenForTeamsUserWithWrongClientIdShouldThrow()
         {
             if (TestEnvironment.ShouldIgnoreIdentityExchangeTokenTest)
             {
                 Assert.Ignore("Ignore exchange teams token test if flag is enabled.");
             }
 
-            var teamsUserParams = await createTeamsUserParams();
             try
             {
                 CommunicationIdentityClient client = CreateClientWithConnectionString();
-                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(teamsUserParams.Token, teamsUserParams.UserId, teamsUserParams.UserId);
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(CTEOptions.TeamsUserAadToken, CTEOptions.UserObjectId, CTEOptions.UserObjectId));
             }
             catch (RequestFailedException ex)
             {
@@ -260,20 +259,19 @@ namespace Azure.Communication.Identity.Tests
         }
 
         [Test]
-        [TestCase("", TestName = "GetTokenForTeamsUserWithEmptyUserIdShouldThrow")]
-        [TestCase("invalid", TestName = "GetTokenForTeamsUserWithInvalidUserIdShouldThrow")]
-        public async Task GetTokenForTeamsUserWithInvalidUserIdShouldThrow(string userId)
+        [TestCase("", TestName = "GetTokenForTeamsUserWithEmptyUserObjectIdShouldThrow")]
+        [TestCase("invalid", TestName = "GetTokenForTeamsUserWithInvalidUserObjectIdShouldThrow")]
+        public async Task GetTokenForTeamsUserWithInvalidUserObjectIdShouldThrow(string userObjectId)
         {
             if (TestEnvironment.ShouldIgnoreIdentityExchangeTokenTest)
             {
                 Assert.Ignore("Ignore exchange teams token test if flag is enabled.");
             }
 
-            var teamsUserParams = await createTeamsUserParams();
             try
             {
                 CommunicationIdentityClient client = CreateClientWithConnectionString();
-                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(teamsUserParams.Token, teamsUserParams.AppId, userId);
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(CTEOptions.TeamsUserAadToken, CTEOptions.ClientId, userObjectId));
             }
             catch (RequestFailedException ex)
             {
@@ -286,18 +284,17 @@ namespace Azure.Communication.Identity.Tests
         }
 
         [Test]
-        public async Task GetTokenForTeamsUserWithWrongUserIdShouldThrow()
+        public async Task GetTokenForTeamsUserWithWrongUserObjectIdShouldThrow()
         {
             if (TestEnvironment.ShouldIgnoreIdentityExchangeTokenTest)
             {
                 Assert.Ignore("Ignore exchange teams token test if flag is enabled.");
             }
 
-            var teamsUserParams = await createTeamsUserParams();
             try
             {
                 CommunicationIdentityClient client = CreateClientWithConnectionString();
-                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(teamsUserParams.Token, teamsUserParams.AppId, teamsUserParams.AppId);
+                Response<AccessToken> tokenResponse = await client.GetTokenForTeamsUserAsync(new GetTokenForTeamsUserOptions(CTEOptions.TeamsUserAadToken, CTEOptions.ClientId, CTEOptions.ClientId));
             }
             catch (RequestFailedException ex)
             {