+++ sdk/communication/Azure.Communication.Identity/CHANGELOG.md
@@ -3,10 +3,10 @@
 ## 1.1.0 (Unreleased)
 
 ### Features Added
-- Added support to build custom Teams endpoint using M365 Teams identities:
-    - Added `GetTokenForTeamsUser(string teamsUserAadToken, string appId, string userId, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken))` method that provides the ability to exchange an AAD access token of a Teams user for a Communication Identity access token to `CommunicationIdentityClient`.
+- Added support to build a custom Teams endpoint using Microsoft 365 Teams identities:
+    - Added `GetTokenForTeamsUser(GetTokenForTeamsUserOptions options, CancellationToken cancellationToken = default)` method that provides the ability to exchange an Azure AD access token of a Teams user for a Communication Identity access token to `CommunicationIdentityClient`.
 - Removed `ServiceVersion.V2021_10_31_preview`
-- Added new API version `ServiceVersion.V2022_06_01` that is default API version
+- Added a new API version `ServiceVersion.V2022_06_01` that is now the default API version
 
 ### Breaking Changes